make Curated the default screen on mobile app
