+++ apps/mobile/src/screens/HomeScreen/LatestScreen.tsx
@@ -1,12 +1,10 @@
-import AsyncStorage from '@react-native-async-storage/async-storage';
-import { Suspense, useCallback, useEffect, useMemo, useState } from 'react';
+import { Suspense, useState } from 'react';
 import { graphql, useFragment, useLazyLoadQuery } from 'react-relay';
 import isFeatureEnabled, { FeatureFlag } from 'src/utils/isFeatureEnabled';
 
 import { FEED_PER_PAGE } from '~/components/Feed/constants';
 import { ActiveFeed } from '~/components/Feed/FeedFilter';
 import { FollowingFeed } from '~/components/Feed/FollowingFeed';
-import { WelcomeToBeta } from '~/components/WelcomeToBeta';
 import { LatestScreenFeatureQuery } from '~/generated/LatestScreenFeatureQuery.graphql';
 import { LatestScreenFragment$key } from '~/generated/LatestScreenFragment.graphql';
 import { LatestScreenQuery } from '~/generated/LatestScreenQuery.graphql';
@@ -24,14 +22,6 @@ function LatestScreenInner({ queryRef }: LatestScreenInnerProps) {
       fragment LatestScreenFragment on Query {
         ...FollowingFeedFragment
         ...WorldwideFeedFragment
-
-        viewer {
-          ... on Viewer {
-            user {
-              username
-            }
-          }
-        }
       }
     `,
     queryRef
@@ -39,33 +29,10 @@ function LatestScreenInner({ queryRef }: LatestScreenInnerProps) {
 
   const [activeFeed, setActiveFeed] = useState<ActiveFeed>('Worldwide');
 
-  const [showWelcome, setShowWelcome] = useState(false);
-
-  const checkShouldShowWelcome = useCallback(async () => {
-    const shown = await AsyncStorage.getItem('welcomeMessageShown');
-    if (shown !== 'true') {
-      setShowWelcome(true);
-      await AsyncStorage.setItem('welcomeMessageShown', 'true');
-    }
-  }, [setShowWelcome]);
-
-  useEffect(() => {
-    checkShouldShowWelcome();
-  }, [checkShouldShowWelcome]);
-
-  const FeedComponent = useMemo(() => {
-    if (activeFeed === 'Following') {
-      return <FollowingFeed queryRef={query} onChangeFeedMode={setActiveFeed} />;
-    } else {
-      return <WorldwideFeed queryRef={query} onChangeFeedMode={setActiveFeed} />;
-    }
-  }, [activeFeed, query]);
-
-  return (
-    <>
-      {FeedComponent}
-      {showWelcome && <WelcomeToBeta username={query.viewer?.user?.username ?? ''} />}
-    </>
+  return activeFeed === 'Following' ? (
+    <FollowingFeed queryRef={query} onChangeFeedMode={setActiveFeed} />
+  ) : (
+    <WorldwideFeed queryRef={query} onChangeFeedMode={setActiveFeed} />
   );
 }
