+++ apps/mobile/src/screens/Login/NotificationUpsellScreen.tsx
@@ -24,7 +24,7 @@ export function NotificationUpsellScreen() {
       routes: [
         {
           name: 'MainTabs',
-          params: { screen: 'HomeTab', params: { screen: 'Home', params: { screen: 'Latest' } } },
+          params: { screen: 'HomeTab', params: { screen: 'Home', params: { screen: 'Curated' } } },
         },
       ],
     });
